--- lib/src/Kaufvertrag/Interfaces/IDataLayer.java
@@ -1,6 +1,0 @@
-package Kaufvertrag.Interfaces;
-
-public interface IDataLayer {
-    IDao<IVertragspartner,String> getDaoVertragspartner();
-    IDao<IWare,Long> getDaoWare();
-}
