--- lib/src/Kaufvertrag/SQL/WareDaoSqlite.java
@@ -1,4 +1,0 @@
-package Kaufvertrag.SQL;
-
-public class WareDaoSqlite {
-}
