--- lib/src/Kaufvertrag/XML/ServiceXml.java
@@ -1,4 +1,0 @@
-package Kaufvertrag.XML;
-
-public class ServiceXml {
-}
