+++ lib/src/Kaufvertrag/dataLayer/businessObjects/Adresse.java
@@ -1,6 +1,6 @@
-package Kaufvertrag;
+package Kaufvertrag.dataLayer.businessObjects;
 
-import Kaufvertrag.Interfaces.IAdresse;
+import Kaufvertrag.businessObjects.IAdresse;
 
 //TODO: find out if Adress implements, extends or smth like that with Adresse (composition)
 public class Adresse implements IAdresse {
