+++ lib/src/Kaufvertrag/dataLayer/businessObjects/dataAccessObjects/DataLayerManager.java
@@ -1,0 +1,6 @@
+package Kaufvertrag.dataLayer.businessObjects.dataAccessObjects;
+
+public class DataLayerManager {
+    
+
+}
