+++ lib/src/Kaufvertrag/dataLayer/businessObjects/dataAccessObjects/XML/DataLayerXml.java
@@ -1,0 +1,20 @@
+package Kaufvertrag.dataLayer.businessObjects.dataAccessObjects.XML;
+
+import Kaufvertrag.dataLayer.businessObjects.dataAccessObjects.IDao;
+import Kaufvertrag.dataLayer.businessObjects.dataAccessObjects.IDataLayer;
+import Kaufvertrag.businessObjects.IVertragspartner;
+import Kaufvertrag.businessObjects.IWare;
+
+public class DataLayerXml implements IDataLayer {
+
+    @Override
+    public IDao<IVertragspartner, String> getDaoVertragspartner() {
+        return null;
+    }
+
+    @Override
+    public IDao<IWare, Long> getDaoWare() {
+        return null;
+    }
+
+}
