+++ lib/src/Kaufvertrag/dataLayer/businessObjects/dataAccessObjects/XML/ServiceXml.java
@@ -1,0 +1,4 @@
+package Kaufvertrag.dataLayer.businessObjects.dataAccessObjects.XML;
+
+public class ServiceXml {
+}
