+++ lib/src/Kaufvertrag/dataLayer/businessObjects/dataAccessObjects/sqlite/WareDaoSqlite.java
@@ -1,0 +1,4 @@
+package Kaufvertrag.dataLayer.businessObjects.dataAccessObjects.sqlite;
+
+public class WareDaoSqlite {
+}
