+++ lib/src/Kaufvertrag/exceptions/DaoException.java
@@ -1,4 +1,4 @@
-package Kaufvertrag.Exceptions;
+package Kaufvertrag.exceptions;
 public class DaoException extends Exception {
     public DaoException(String message){
         message = super.getMessage();
